Refactor : move testutils from `exoscale` to `utils` to avoid import cycles
