+++ pkg/resources/anti_affinity_group/resource.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 
 	"github.com/exoscale/terraform-provider-exoscale/pkg/config"
-	"github.com/exoscale/terraform-provider-exoscale/pkg/utils"
+	"github.com/exoscale/terraform-provider-exoscale/utils"
 
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
