+++ pkg/resources/database/resource_user.go
@@ -17,7 +17,7 @@ import (
 
 	exoscale "github.com/exoscale/egoscale/v3"
 	"github.com/exoscale/terraform-provider-exoscale/pkg/config"
-	"github.com/exoscale/terraform-provider-exoscale/pkg/utils"
+	"github.com/exoscale/terraform-provider-exoscale/utils"
 )
 
 // UserResource defines the resource implementation.
