+++ pkg/resources/instance/datasource_list.go
@@ -16,7 +16,7 @@ import (
 
 	"github.com/exoscale/terraform-provider-exoscale/pkg/config"
 	"github.com/exoscale/terraform-provider-exoscale/pkg/filter"
-	"github.com/exoscale/terraform-provider-exoscale/pkg/utils"
+	"github.com/exoscale/terraform-provider-exoscale/utils"
 )
 
 func DataSourceList() *schema.Resource {
