+++ pkg/resources/instance_pool/datasource.go
@@ -11,7 +11,7 @@ import (
 
 	v3 "github.com/exoscale/egoscale/v3"
 	"github.com/exoscale/terraform-provider-exoscale/pkg/config"
-	"github.com/exoscale/terraform-provider-exoscale/pkg/utils"
+	"github.com/exoscale/terraform-provider-exoscale/utils"
 )
 
 // DataSourceSchema returns a schema for a single instance pool data source.
