+++ pkg/resources/instance_pool/datasource_list.go
@@ -14,7 +14,7 @@ import (
 	v3 "github.com/exoscale/egoscale/v3"
 
 	"github.com/exoscale/terraform-provider-exoscale/pkg/config"
-	"github.com/exoscale/terraform-provider-exoscale/pkg/utils"
+	"github.com/exoscale/terraform-provider-exoscale/utils"
 )
 
 func DataSourceList() *schema.Resource {
